process: Encode hash value to map AFL bitmap
This commit adds encoding hash value to map AFL bitmap index to be
updated. It is hard to figure out what encoding algorithm should be used
because the reference code differs from the proposed method in the
paper. However, what is the most important point is that the encoding
has uniform distribution, so we xorshift pseudorandom number generator
for encoding.

Signed-off-by: Akira Moroo <retrage01@gmail.com>